+++ internal/provider/provider.go
@@ -259,6 +259,7 @@ func (p *SdwanProvider) Resources(ctx context.Context) []func() resource.Resourc
 		NewCLIDeviceTemplateResource,
 		NewFeatureDeviceTemplateResource,
 		NewAttachFeatureDeviceTemplateResource,
+		NewLocalizedPolicyResource,
 	}
 }
 
@@ -306,6 +307,7 @@ func (p *SdwanProvider) DataSources(ctx context.Context) []func() datasource.Dat
 		NewRoutePolicyDefinitionDataSource,
 		NewCLIDeviceTemplateDataSource,
 		NewFeatureDeviceTemplateDataSource,
+		NewLocalizedPolicyDataSource,
 	}
 }