+++ docs/guides/changelog.md
@@ -15,6 +15,7 @@ description: |-
 - Add `sdwan_acl_policy_definition` resource and data source
 - Add `sdwan_device_acl_policy_definition` resource and data source
 - Add `sdwan_route_policy_definition` resource and data source
+- Add `sdwan_localized_policy` resource and data source
 
 ## 0.1.6